- Fix bug where spell level select could show as a level the caster did not have - Change the way portrait talking timing is handled - Selecting spells from the menu now show range and area - Resources now load in the background in development mode - "BulkLoaded" resources are now broken down in to their actually items so that the load item count is more accurate - More work to abstract engine states to make it mroe configurable - Fix broken animation viewer - Added battle options menu and support for each of the options

+++ CommRPG/src/mb/jython/GlobalPythonFactory.java
@@ -89,10 +89,14 @@ public class GlobalPythonFactory
 			JythonObjectFactory.sys.path.append(new PyString(JythonObjectFactory.sys.getPath("scripts")));
 
 			File scriptsFolder = new File(JythonObjectFactory.sys.getPath("scripts"));
-			for (File file : scriptsFolder.listFiles())
-			{
-				if (file.getName().endsWith(".class"))
-					file.delete();
+			if (scriptsFolder.exists()) {
+				for (File file : scriptsFolder.listFiles())
+				{
+					if (file.getName().endsWith(".class"))
+						file.delete();
+				}
+			} else {
+				return;
 			}
 		}
